new repl view
